requirements.txt
flask
python-dotenv
requests
chromadb
sentence-transformers
PyMuPDF